refactor: renaming assets (#138)
* feat: open transaction link in new tab for convenience

* fix: renmae assets folder
