+++ src/components/Withdrawal/WithdrawalItemV0.tsx
@@ -5,9 +5,9 @@ import { getDisplayAmount } from "../../utils/formatTokenAmount";
 import { BI, Cell, HexNumber, HexString } from "@ckb-lumos/lumos";
 import { ProxyERC20 } from "../../light-godwoken/lightGodwokenType";
 import { useLightGodwoken } from "../../hooks/useLightGodwoken";
-import { ReactComponent as CKBIcon } from "../../asserts/ckb.svg";
-import { ReactComponent as ArrowDownIcon } from "../../asserts/arrow-down.svg";
-import { ReactComponent as ArrowUpIcon } from "../../asserts/arrow-up.svg";
+import { ReactComponent as CKBIcon } from "../../assets/ckb.svg";
+import { ReactComponent as ArrowDownIcon } from "../../assets/arrow-down.svg";
+import { ReactComponent as ArrowUpIcon } from "../../assets/arrow-up.svg";
 import { MainText } from "../../style/common";
 import { COLOR } from "../../style/variables";
 import { CheckCircleOutlined, CloseCircleOutlined, FieldTimeOutlined } from "@ant-design/icons";
